backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
PyJWT==2.8.0
gunicorn==21.2.0
